Build in-memory DOAs for entities, for use until a persistence DB is available.
Add the matching Engine, with its required Book and Level.  This
Engine basically works, but has no tests.  It implements price-time
priority matching.  Trades are executed at the price of the new order.

Add cancel-replace functionality, with controls on how the new order
should be created when the original ordered has been partially or
fully filled.  Specifying the capping and replacement behavior avoids
issues with race conditions.

Wire up the generators and DAOs, with 100 instruments created on
startup.

Modify the sample python scripts to take arguments for all parameters.