Change timestamp to createTime to avoid collisions with SQL reserved words.
Switch to JdbcClient so that we can easily use named parameters.

Make the database connections fully operational, with a sharded
datbase for Orders, sharded by Instrument (modulo of instrumentId)

Restructure the relationships between OrderState, Trade and TradeLeg
so that OrderState and Trade both have collections of TradeLegs, and
each TradeLeg points to its related OrderState and Trade.

Cache both retrieved and newly submitted Orders.

On startup, load open Orders and OrderStates into the Books.  Do not
load the TradeLegs, but sum their quantities to set the correct
remaining quantity on the OrderStates.

Save Orders, OrderState, Trades and TradeLegs.