Add retrieval of instruments from the database, instead of the previous hard-coded instruments.
Add additional detail to the instruments.

Fetch all instruments when responding to the API request for
instruments, so that instruments which are INACTIVE are returned,
allowing the client to mark their own copy as inactive.

However, when starting OpenExchange, only use ACTIVE instruments for
starting engines.

Fix some Postgres schema creation defects.